database/requirements.txt
psycopg2-binary >=2.9.3
schedule >= 1.1.0